requirements.txt
backcall==0.2.0
colorama==0.4.5
cycler==0.11.0
debugpy==1.6.3
decorator==5.1.1
entrypoints==0.4
fonttools==4.38.0
ipykernel==6.16.0
ipython==7.34.0
jedi==0.18.1
jupyter-core==4.11.1
jupyter_client==7.4.2
kiwisolver==1.4.4
matplotlib==3.5.3
matplotlib-inline==0.1.6
nest-asyncio==1.5.6
numpy==1.21.6
packaging==21.3
pandas==1.3.5
parso==0.8.3
pickleshare==0.7.5
Pillow==9.3.0
prompt-toolkit==3.0.31
psutil==5.9.2
Pygments==2.13.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.5
pywin32==304
pyzmq==24.0.1
scipy==1.7.3
seabreeze==2.0.3
six==1.16.0
tornado==6.2
traitlets==5.4.0
typing_extensions==4.4.0
wcwidth==0.2.5
